move examples test file to test directory
Rename run functions to main. Use pathlib.Path more.
